refactor: move newton stuff to its own file
